Add number to choose wheight code for recieving messages from iothub and iothub eventhub

+++ Bound.Tablet/Bound.Tablet/Models/DeviceData.cs
@@ -9,7 +9,7 @@ namespace Bound
         public string MachineName { get; set; }
         public string ObjectId { get; set; }
 
-        public int Weight { get; set; }
+        public long Weight { get; set; }
 
         public IoTHubDevice Device { get; set; } = new IoTHubDevice();
     }
